att_from_rscripts {attachment}	R Documentation

Look for functions called with :: and library/requires in folder of scripts

Description

Look for functions called with :: and library/requires in folder of scripts

Usage

att_from_rscripts(
  path = "R",
  pattern = "*.[.](r|R)$",
  recursive = TRUE,
  folder_to_exclude = "renv"
)

Arguments

path
directory with R scripts inside or vector of R scripts

pattern
pattern to detect R script files

recursive
logical. Should the listing recurse into directories?

folder_to_exclude
Folder to exclude during scan to detect packages. 'renv' by default.

Value

vector of character of packages names found in the R script

Examples

dummypackage <- system.file("dummypackage",package = "attachment")
# browseURL(dummypackage)

att_from_rscripts(path = file.path(dummypackage, "R"))
att_from_rscripts(path = list.files(file.path(dummypackage, "R"), full.names = TRUE))

[Package attachment version 0.4.5 Index]