attachment-package {attachment}	R Documentation

attachment: Deal with Dependencies

Description

Manage dependencies during package development. This can retrieve all dependencies that are used in ".R" files in the "R/" directory, in ".Rmd" files in "vignettes/" directory and in 'roxygen2' documentation of functions. There is a function to update the "DESCRIPTION" file of your package with 'CRAN' packages or any other remote package. All functions to retrieve dependencies of ".R" scripts and ".Rmd" or ".qmd" files can be used independently of a package development.

Author(s)

Maintainer: Vincent Guyader vincent@thinkr.fr (ORCID)

Authors:

Sébastien Rochette sebastienrochettefr@gmail.com (ORCID) (previous maintainer)

Murielle Delmotte murielle@thinkr.fr (ORCID)

Swann Floc'hlay swann@thinkr.fr (ORCID)

Other contributors:

ThinkR [copyright holder, funder]

See Also

Useful links:

https://thinkr-open.github.io/attachment/

https://github.com/ThinkR-open/attachment

Report bugs at https://github.com/ThinkR-open/attachment/issues

[Package attachment version 0.4.5 Index]