att_from_examples {attachment}	R Documentation

Get all packages called in examples from R files

Description

Get all packages called in examples from R files

Usage

att_from_examples(dir.r = "R")

Arguments

dir.r
path to directory with R scripts.

Value

Character vector of packages called with library or require.

Examples

dummypackage <- system.file("dummypackage",package = "attachment")

# browseURL(dummypackage)
att_from_examples(dir.r = file.path(dummypackage,"R"))

[Package attachment version 0.4.5 Index]