attachment: Deal with Dependencies

Manage dependencies during package development. This can retrieve all dependencies that are used in ".R" files in the "R/" directory, in ".Rmd" files in "vignettes/" directory and in 'roxygen2' documentation of functions. There is a function to update the "DESCRIPTION" file of your package with 'CRAN' packages or any other remote package. All functions to retrieve dependencies of ".R" scripts and ".Rmd" or ".qmd" files can be used independently of a package development.

See also

Author

Maintainer: Sébastien Rochette sebastien@thinkr.fr (ORCID)

Authors:

Vincent Guyader vincent@thinkr.fr (ORCID) (previous maintainer)

Murielle Delmotte murielle@thinkr.fr (ORCID)

Swann Floc'hlay swann@thinkr.fr (ORCID)

Other contributors:

ThinkR [copyright holder, funder]